requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0

# Async HTTP
httpx==0.26.0
aiofiles==23.2.1

# Redis
redis[hiredis]==5.0.1

# Database
asyncpg==0.29.0

# AI SDKs
anthropic==0.19.1
openai==1.12.0
google-generativeai==0.3.2

# GitHub
PyJWT==2.8.0
cryptography==42.0.2

# Observability
sentry-sdk==1.40.5
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
tenacity==8.2.3
python-multipart==0.0.9

# Development
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
ruff==0.2.1
mypy==1.8.0